3.26 backup

+++ src/main/java/cc/aisc/ybk/content/service/UserService.java
@@ -6,6 +6,7 @@ import cc.aisc.ybk.content.model.User;
  * Created by sjf on 16-3-19.
  */
 public interface UserService {
-    User findById(Long id);
+
+    User findById(Integer id);
     int create(User user);
 }
